+++ src/main/java/site/gachontable/domain/member/exception/UserNotFoundException.java
@@ -1,0 +1,10 @@
+package site.gachontable.domain.member.exception;
+
+import site.gachontable.independent.type.ErrorCode;
+import site.gachontable.independent.error.exception.ServiceException;
+
+public class UserNotFoundException extends ServiceException {
+    public UserNotFoundException() {
+        super(ErrorCode.USER_NOT_FOUND);
+    }
+}
