+++ src/main/java/site/gachontable/domain/member/port/in/UserLogin.java
@@ -1,0 +1,8 @@
+package site.gachontable.domain.member.port.in;
+
+import site.gachontable.infra.security.jwt.dto.JwtResponse;
+
+public interface UserLogin {
+
+    JwtResponse execute(String id, String password);
+}
