+++ src/main/java/site/gachontable/domain/member/port/in/UserRegister.java
@@ -1,0 +1,8 @@
+package site.gachontable.domain.member.port.in;
+
+import site.gachontable.domain.shared.dto.response.RegisterResponse;
+
+public interface UserRegister {
+
+    RegisterResponse execute(String username, String password, String tel);
+}
