--- src/main/java/site/gachontable/gachontablebe/domain/user/usecase/UserLogin.java
@@ -1,8 +1,0 @@
-package site.gachontable.gachontablebe.domain.user.usecase;
-
-import site.gachontable.gachontablebe.global.jwt.dto.JwtResponse;
-
-public interface UserLogin {
-
-    JwtResponse execute(String id, String password);
-}
